update: Add org.jsonurl.j2se Test package
Now that the org.jsonurl.j2se package has been created for the code it
makes sense to separate the tests as well.
