+++ module/jsonurl-core/src/test/java/org/jsonurl/j2se/PrimitiveParseTest.java
@@ -1,6 +1,4 @@
-package org.jsonurl;
-
-import org.jsonurl.j2se.JavaValueFactory;
+package org.jsonurl.j2se;
 
 /*
  * Copyright 2019 David MacCormack
@@ -20,18 +18,18 @@ import org.jsonurl.j2se.JavaValueFactory;
  */
 
 /**
- * Unit test using JavaValueFactory.BIGMATH
+ * Unit test using JavaValueFactory.PRIMITIVE
  *
  * @author jsonurl.org
  * @author David MacCormack
  * @since 2019-09-01
  */
-public class BigMathParseTest extends JavaValueFactoryParseTest {
+public class PrimitiveParseTest extends JavaValueFactoryParseTest {
 
     /**
-     * Create a new BigMathParseTest.
+     * Create a new PrimitiveParseTest.
      */
-    public BigMathParseTest() {
-        super(JavaValueFactory.BIGMATH);
+    public PrimitiveParseTest() {
+        super(JavaValueFactory.PRIMITIVE);
     }
 }
